The ITEM_COUNT counter is just not tracking correctly.  The technique to increment and decrement for various events hasn't worked yet.  It also made me question if it was the right way to do it.  I think if the cache implements Count then that is the way to poll a current count.  With the FusionCache.EventCounter I am committing some experimental work.
Friction Points:
  - IMemoryCache does not have a Count property.  I  have to Inject my own MemoryCache.
  - AddFusionCache extension method does not have a IMemoryCache parameter.  Not a big deal
  - Some usage pattern I have in a project involve some undesirable singletons that store IFusionMetrics and now I would have to take on the IMemoryCache also.  Kind of stuck with this pattern for a while.  Modernizing takes time :(.

The experiment results in accurate counts.  Still need to unwire the old way of tracking but this all brings me to a change to the way I inject IFusionMetrics.

My provider pattern of passing IFusionMetrics would be better implemented with a Plugin that is loaded based on assembly scanning at startup.  Then IFusionMetrics and IFusionCache wouldn’t both have to be passed an IMemoryCache.  The FusionCache implementation would give IFusionMetrics the IMemoryCache.   The implementation of IFusionMetrics would code for the existence of a Count property implementation.